In fut_set implement hash for tuples.
