+++ src/CMakeLists.txt
@@ -19,5 +19,6 @@ install (FILES pdaaal/vector_set.h pdaaal/fut_set.h pdaaal/NFA.h pdaaal/Weight.h
         pdaaal/PDAFactory.h pdaaal/SolverInstance.h
         pdaaal/ParsingPDAFactory.h
         pdaaal/Refinement.h
+        pdaaal/std20.h
         pdaaal/AbstractionMapping.h pdaaal/AbstractionPDA.h pdaaal/AbstractionPAutomaton.h pdaaal/CegarPdaFactory.h pdaaal/ptrie_interface.h
         pdaaal/SimplePDAFactory.h pdaaal/TypedPDA.h pdaaal/PAutomaton.h pdaaal/Solver.h pdaaal/Reducer.h DESTINATION include/pdaaal)
