+++ src/pdaaal/fut_set.h
@@ -28,30 +28,8 @@
 #define PDAAAL_FUT_SET_H
 
 #include "vector_set.h"
-
-namespace std20 { // TODO: When C++20 arrives: Delete this.
-    // Add contains method to unordered containers, until we can use the ones in C++20.
-    template<typename Key, typename Tp,
-            typename Hash = std::hash<Key>,
-            typename Pred = std::equal_to<Key>,
-            typename Alloc = std::allocator<std::pair<const Key, Tp>>>
-    class unordered_map : public std::unordered_map<Key,Tp,Hash,Pred,Alloc> {
-    public:
-        bool contains(const Key &key) const {
-            return this->find(key) != this->end();
-        }
-    };
-    template<typename Value,
-            typename Hash = std::hash<Value>,
-            typename Pred = std::equal_to<Value>,
-            typename Alloc = std::allocator<Value>>
-    class unordered_set : public std::unordered_set<Value,Hash,Pred,Alloc> {
-    public:
-        bool contains(const Value &value) const {
-            return this->find(value) != this->end();
-        }
-    };
-}
+#include <boost/functional/hash.hpp>
+#include "std20.h"
 
 namespace pdaaal::fut {
 
@@ -61,14 +39,31 @@ namespace pdaaal::fut {
     };
 
     namespace detail {
+
+        // Use std::hash by default, but for tuples use custom implementation based on boost::hash_combine, but using std::hash on inner types.
+        template <typename T>
+        struct hash {
+            size_t operator()(const T& t) const {
+                return std::hash<T>()(t);
+            }
+        };
+        template <typename... Args>
+        struct hash<std::tuple<Args...>> {
+            size_t operator()(const std::tuple<Args...>& tuple) const {
+                size_t seed = 0;
+                std::apply([&seed](auto&&... args){ (boost::hash_detail::hash_combine_impl(seed, std::hash<std20::remove_cvref_t<decltype(args)>>()(args)), ...); }, tuple);
+                return seed;
+            }
+        };
+
         // FUT-set is a Fast Unordered Tuple-set. Hopefully so fast that executing it says FUT.
         template<class T, type... C>
         class fut_set { };
 
         template<typename Key, typename Value, type C>
-        using map_container = std::conditional_t<C == type::hash, std20::unordered_map<Key, Value>, vector_map<Key, Value>>;
+        using map_container = std::conditional_t<C == type::hash, std20::unordered_map<Key, Value, hash<Key>>, vector_map<Key, Value>>;
         template<typename Key, type C>
-        using set_container = std::conditional_t<C == type::hash, std20::unordered_set<Key>, vector_set<Key>>;
+        using set_container = std::conditional_t<C == type::hash, std20::unordered_set<Key, hash<Key>>, vector_set<Key>>;
 
         // TODO: Add 'compare_by' functionality somehow.
