+++ src/pdaaal/ptrie_interface.h
@@ -29,15 +29,7 @@
 
 #include <ptrie/ptrie_map.h>
 #include <vector>
-
-namespace std20{
-    template< class T >
-    struct remove_cvref {
-        typedef std::remove_cv_t<std::remove_reference_t<T>> type;
-    };
-    template< class T >
-    using remove_cvref_t = typename remove_cvref<T>::type;
-}
+#include "std20.h"
 
 namespace pdaaal::utils {
     // This file defines interfacing functionality for using more types with ptrie.
